refactor: extract send_email_message function to handle single email sending
