+++ django_nitro_mailer/tasks.py
@@ -20,6 +20,34 @@ def throttle_email_delivery() -> None:
         time.sleep(throttle_delay / 1000)
 
 
+def send_email_message(email_obj: Email, connection) -> bool:
+    try:
+        email_message = email_obj.email
+        if email_message:
+            connection.send_messages([email_message])
+
+            if email_db_logging:
+                EmailLog.objects.create(email_data=email_obj.email_data, result=EmailLog.Results.SUCCESS)
+
+            logger.info(
+                "Email sent successfully",
+                extra={"recipients": email_obj.recipients, "created_at": timezone.now()},
+            )
+
+            email_obj.delete()
+
+            return True
+        else:
+            logger.error("Failed to retrieve email")
+            return False
+    except Exception as e:
+        if email_db_logging:
+            EmailLog.objects.create(email_data=email_obj.email_data, result=EmailLog.Results.FAILURE)
+
+        logger.error("Failed to send email", exc_info=e)
+        return False
+
+
 def send_emails(queryset: Optional[models.QuerySet] = None) -> None:
     if queryset is None:
         queryset = Email.objects.exclude(priority=Email.Priorities.DEFERRED).order_by("-priority", "created_at")
@@ -28,28 +56,5 @@ def send_emails(queryset: Optional[models.QuerySet] = None) -> None:
 
     with transaction.atomic():
         for email_obj in queryset.select_for_update(nowait=True):
-            try:
-                email_message = email_obj.email
-                if email_message:
-                    connection.send_messages([email_message])
-
-                    if email_db_logging:
-                        EmailLog.objects.create(email_data=email_obj.email_data, result=EmailLog.Results.SUCCESS)
-
-                    logger.info(
-                        "Email sent successfully",
-                        extra={"recipients": email_obj.recipients, "created_at": timezone.now()},
-                    )
-
-                    email_obj.delete()
-
-                else:
-                    logger.error("Failed to retrieve email")
-                throttle_email_delivery()
-            except Exception as e:
-                if email_db_logging:
-                    EmailLog.objects.create(email_data=email_obj.email_data, result=EmailLog.Results.FAILURE)
-
-                logger.error("Failed to send email", exc_info=e)
-
+            send_email_message(email_obj, connection)
             throttle_email_delivery()
